// src/app/model/employee.model.ts
export interface Employee {
  employee_id: string;
  employee_name: string;
  employee_en: string;
  employee_extension: string;
  employee_mobile: string;
  employee_address: string;
  employee_position: string;
  employee_status: string;
  employee_email: string;
  employee_password: string;
  employee_bd: Date;
  employee_onboarddate: Date;
}

export interface LoginInfo {
  employee_email: string;
  employee_password: string;
}